Code cleaning + errors catching

+++ phpoole.php
@@ -10,6 +10,8 @@
  * Copyright (c) 2013 Arnaud Ligny
  */
 
+error_reporting(0); 
+
 use Zend\Loader\StandardAutoloader;
 use Zend\Console;
 use Michelf\Markdown;
@@ -18,20 +20,9 @@ use Michelf\Markdown;
 if (file_exists('vendor/autoload.php')) {
     $loader = include 'vendor/autoload.php';
 }
-// Includes ZF2 components
-if (is_dir('vendor/zendframework')) {
-    $zf2Path = 'vendor/zendframework';
-    if (isset($loader)) {
-        $loader->add('Zend', $zf2Path);
-    }
-}
-if (!class_exists('Zend\Loader\AutoloaderFactory')) {
-    echo 'Unable to load ZF2 components. Run `php composer.phar install`.';
-    exit(2);
-}
 
-$pwd = getcwd();
-$websiteDir = null;
+define('PHPOOLE_DIRNAME', '_phpoole');
+$websitePath = getcwd();
 
 // Defines rules
 $rules = array(
@@ -60,71 +51,41 @@ if ($opts->getOption('help')) {
 
 // init option
 if ($opts->getOption('init')) {
-    if (isset($opts->i)) {
-        if (!is_dir($opts->i)) {
-            echo 'Invalid directory provided' . PHP_EOL
-                . PHP_EOL;
-            echo $opts->getUsageMessage();
-            exit(2);
-        }
-        $websiteDir = $opts->i;
-        $websiteDir = str_replace('\\', '/', realpath($websiteDir));
-    }
-    echo 'Initializing new website' . (!is_null($websiteDir) ? " in $websiteDir" : '') . PHP_EOL
-        . PHP_EOL;
-    Init((!is_null($websiteDir) ? $websiteDir : $pwd));
-    echo "[OK] create .phpoole" . PHP_EOL;
-    MakeConfigFile((!is_null($websiteDir) ? $websiteDir : $pwd) . '/.phpoole/config.ini');
-    echo "[OK] create .phpoole/router.php" . PHP_EOL;
-    MakeRouterFile((!is_null($websiteDir) ? $websiteDir : $pwd) . '/router.php');
-    echo '[OK] create .phpoole/config.ini' . PHP_EOL;
-    mkdir((!is_null($websiteDir) ? $websiteDir : $pwd) . '/.phpoole/layouts');
-    echo '[OK] create .phpoole/layouts' . PHP_EOL;
-    MakeLayoutBaseFile((!is_null($websiteDir) ? $websiteDir : $pwd) . '/.phpoole/layouts/base.html');
-    echo '[OK] create .phpoole/layouts/base.html' . PHP_EOL;
-    MakeAssetsDir((!is_null($websiteDir) ? $websiteDir : $pwd) . '/.phpoole');
-    echo '[OK] create .phpoole/assets/...' . PHP_EOL;
-    MakeContentDir((!is_null($websiteDir) ? $websiteDir : $pwd) . '/.phpoole');
-    echo '[OK] create .phpoole/content/...' . PHP_EOL;
-    MakeIndexFile((!is_null($websiteDir) ? $websiteDir : $pwd) . '/.phpoole/content/pages/index.md');
-    echo '[OK] create .phpoole/content/pages/index.md' . PHP_EOL;
+    $websitePath = getOptionWebsitePath($opts, 'i');
+    printf('Initializing new website in %s' . PHP_EOL . PHP_EOL, $websitePath);
+    mkInitDir($websitePath);
+    mkConfigFile($websitePath . '/' . PHPOOLE_DIRNAME . '/config.ini');
+    mkLayoutsDir($websitePath . '/' . PHPOOLE_DIRNAME);
+    mkLayoutBaseFile($websitePath . '/' . PHPOOLE_DIRNAME, 'base.html');
+    mkAssetsDir($websitePath . '/' . PHPOOLE_DIRNAME);
+    mkContentDir($websitePath . '/'. PHPOOLE_DIRNAME);
+    mkContentIndexFile($websitePath . '/' . PHPOOLE_DIRNAME, 'index.md');
+    mkRouterFile($websitePath . '/router.php');
     exit(0);
 }
 
 // generate option
 if ($opts->getOption('generate')) {
-    if (isset($opts->g)) {
-        if (!is_dir($opts->g)) {
-            echo 'Invalid directory provided' . PHP_EOL
-                . PHP_EOL;
-            echo $opts->getUsageMessage();
-            exit(2);
-        }
-        $websiteDir = $opts->g;
-        $websiteDir = str_replace('\\', '/', realpath($websiteDir));
-    }
-    echo 'Generating website' . (!is_null($websiteDir) ? " in $websiteDir" : '') . PHP_EOL
-        . PHP_EOL;
-    if (!is_dir((!is_null($websiteDir) ? $websiteDir : $pwd) . '/.phpoole/layouts')) {
-        echo 'Initiate website before try to generate' . PHP_EOL
-            . PHP_EOL;
-        echo $opts->getUsageMessage();
-        exit(2);
-    }
-    $twigLoader = new Twig_Loader_Filesystem((!is_null($websiteDir) ? $websiteDir : $pwd) . '/.phpoole/layouts');
+    $websitePath = getOptionWebsitePath($opts, 'g');
+    printf('Generating website in %s' . PHP_EOL . PHP_EOL, $websitePath);
+    $twigLoader = new Twig_Loader_Filesystem($websitePath . '/' . PHPOOLE_DIRNAME . '/layouts');
     $twig = new Twig_Environment($twigLoader, array('autoescape' => false));
-    $pagesPath = (!is_null($websiteDir) ? $websiteDir : $pwd) . '/.phpoole/content/pages';
-    $iterator = new MarkdownFileFilter($pagesPath);
-    foreach ($iterator as $filePage) {
+    $pagesPath = $websitePath . '/' . PHPOOLE_DIRNAME . '/content/pages';
+    $markdownIterator = new MarkdownFileFilter($pagesPath);
+    foreach ($markdownIterator as $filePage) {
         $page = parseContent(
             file_get_contents($filePage->getPathname()),
             $filePage->getFilename()
         );
+        if (isset($page['title'])) {
+            $title = $page['title'];
+        }
+        else {
+            $title = 'PHPoole static website';
+        }
         if (
             isset($page['layout'])
-            && is_file(
-                (!is_null($websiteDir) ? $websiteDir : $pwd) . '/.phpoole/layouts' . '/' . $page['layout'] . '.html'
-            )
+            && is_file($websitePath . '/' . PHPOOLE_DIRNAME . '/layouts' . '/' . $page['layout'] . '.html')
         ) {
             $layout = $page['layout'] . '.html';
         }
@@ -132,48 +93,50 @@ if ($opts->getOption('generate')) {
             $layout = 'base.html';
         }
         $rendered = $twig->render($layout, array(
-            'title'   => $page['title'],
+            'title'   => $title,
             'content' => $page['body']
         ));
-        if (!is_dir((!is_null($websiteDir) ? $websiteDir : $pwd) . '/' . $iterator->getSubPath())) {
-            mkdir((!is_null($websiteDir) ? $websiteDir : $pwd) . '/' . $iterator->getSubPath(), 0777, true);
+        try {
+            if (!is_dir($websitePath . '/' . $markdownIterator->getSubPath())) {
+                if (false === mkdir($websitePath . '/' . $markdownIterator->getSubPath(), 0777, true)) {
+                    throw new Exception(sprintf('%s not created', $websitePath . '/' . $markdownIterator->getSubPath()));
+                }
+            }
+            if (is_file($websitePath . '/' . ($markdownIterator->getSubPath() != '' ? $markdownIterator->getSubPath() . '/' : '') . $filePage->getBasename('.md') . '.html')) {
+                if (false === unlink($websitePath . '/' . ($markdownIterator->getSubPath() != '' ? $markdownIterator->getSubPath() . '/' : '') . $filePage->getBasename('.md') . '.html')) {
+                    throw new Exception(sprintf('%s%s.html not deleted', ($markdownIterator->getSubPath() != '' ? $markdownIterator->getSubPath() . '/' : ''), $filePage->getBasename('.md')));
+                }
+                echo '[OK] delete ' . ($markdownIterator->getSubPath() != '' ? $markdownIterator->getSubPath() . '/' : '') . $filePage->getBasename('.md') . '.html' . PHP_EOL;
+            }
+            if (false === file_put_contents(sprintf('%s%s.html', $websitePath . '/' . ($markdownIterator->getSubPath() != '' ? $markdownIterator->getSubPath() . '/' : ''), $filePage->getBasename('.md')), $rendered)) {
+                throw new Exception(sprintf('%s%s.html not written', ($markdownIterator->getSubPath() != '' ? $markdownIterator->getSubPath() . '/' : ''), $filePage->getBasename('.md')));
+            }
+            printf('[OK] write %s%s.html' . PHP_EOL, ($markdownIterator->getSubPath() != '' ? $markdownIterator->getSubPath() . '/' : ''), $filePage->getBasename('.md'));
         }
-        if (is_file((!is_null($websiteDir) ? $websiteDir : $pwd) . '/' . ($iterator->getSubPath() != '' ? $iterator->getSubPath() . '/' : '') . $filePage->getBasename('.md') . '.html')) {
-            unlink((!is_null($websiteDir) ? $websiteDir : $pwd) . '/' . ($iterator->getSubPath() != '' ? $iterator->getSubPath() . '/' : '') . $filePage->getBasename('.md') . '.html');
-            echo '[OK] delete ' . ($iterator->getSubPath() != '' ? $iterator->getSubPath() . '/' : '') . $filePage->getBasename('.md') . '.html' . PHP_EOL;
+        catch (Exception $e) {
+            printf('[KO] %s' . PHP_EOL, $e->getMessage());
+            exit(2);
         }
-        file_put_contents(
-            (!is_null($websiteDir) ? $websiteDir : $pwd) . '/' . ($iterator->getSubPath() != '' ? $iterator->getSubPath() . '/' : '') . $filePage->getBasename('.md') . '.html', $rendered
-        );
-        echo '[OK] write ' . ($iterator->getSubPath() != '' ? $iterator->getSubPath() . '/' : '') . $filePage->getBasename('.md') . '.html' . PHP_EOL;
     }
     exit(0);
 }
 
-// Serve option
+// serve option
 if ($opts->getOption('serve')) {
-    if (isset($opts->s)) {
-        if (!is_dir($opts->s)) {
-            echo 'Invalid directory provided' . PHP_EOL
-                . PHP_EOL;
-            echo $opts->getUsageMessage();
-            exit(2);
-        }
-        $websiteDir = $opts->s;
-        $websiteDir = str_replace('\\', '/', realpath($websiteDir));
-    }
-    printf(
-        'Start server http://%s:%d' . PHP_EOL,
-        'localhost',
-        '8000'
-    );
+    $websitePath = getOptionWebsitePath($opts, 's');
+    if (version_compare(PHP_VERSION, '5.4.0', '<')) {
+        echo 'PHP 5.4+ required to run built-in server (your version: ' . PHP_VERSION . ')' . PHP_EOL;
+        exit(2);
+    }    
+    printf('Start server http://%s:%d' . PHP_EOL, 'localhost', '8000');
+    echo 'Ctrl+C to stop it.' . PHP_EOL;
     if (strtoupper(substr(PHP_OS, 0, 3)) === 'WIN') {
         $command = sprintf(
             'START /B php -S %s:%d -t %s %s > nul',
             'localhost',
             '8000',
-            $websiteDir,
-            "$websiteDir/router.php"
+            $websitePath,
+            "$websitePath/router.php"
         );
     }
     else {
@@ -182,46 +145,39 @@ if ($opts->getOption('serve')) {
             'php -S %s:%d -t %s %s >/dev/null',
             'localhost',
             '8000',
-            $websiteDir,
-            "$websiteDir/router.php"
+            $websitePath,
+            "$websitePath/router.php"
         );
     }
-    $output = array(); 
     exec($command);
     exit(0);
 }
 
-// Deploy option
+// deploy option
 if ($opts->getOption('deploy')) {
+    $websitePath = getOptionWebsitePath($opts, 'd');
     echo 'Not yet implemented' . PHP_EOL
         . PHP_EOL;
     exit(0);
 }
 
-// List option
+// list option
 if ($opts->getOption('list')) {
-    if (isset($opts->l)) {
-        if (!is_dir($opts->l)) {
-            echo 'Invalid directory provided' . PHP_EOL
-                . PHP_EOL;
-            echo $opts->getUsageMessage();
-            exit(2);
-        }
-        $websiteDir = $opts->l;
-        $websiteDir = str_replace('\\', '/', realpath($websiteDir));
-    }
-    echo 'List pages' . (!is_null($websiteDir) ? " in $websiteDir" : '') . PHP_EOL
-        . PHP_EOL;
-    if (!is_dir($websiteDir . '/.phpoole/content/pages')) {
+    $websitePath = getOptionWebsitePath($opts, 'l');
+    printf('List content in %s' . PHP_EOL . PHP_EOL, $websitePath);
+    if (!is_dir($websitePath . '/' . PHPOOLE_DIRNAME . '/content/pages')) {
         echo 'Invalid content/pages directory' . PHP_EOL
             . PHP_EOL;
         echo $opts->getUsageMessage();
         exit(2);
     }
-    $fs = new FilesystemIterator($websiteDir . '/.phpoole/content/pages');
-    foreach($fs as $file) {
+    $contentIterator = new RecursiveIteratorIterator(
+        new RecursiveDirectoryIterator($websitePath . '/' . PHPOOLE_DIRNAME . '/content/pages'),
+        RecursiveIteratorIterator::CHILD_FIRST
+    );    
+    foreach($contentIterator as $file) {
         if ($file->isFile()) {
-            echo '- ' . $file->getFilename() . PHP_EOL;
+            printf('- %s%s' . PHP_EOL, ($contentIterator->getSubPath() != '' ? $contentIterator->getSubPath() . '/' : ''), $file->getFilename());
         }
     }
     exit(0);
@@ -236,33 +192,81 @@ exit(2);
  * PHPoole helpers
  */
 
-function Init($path) {
-    if (is_dir($path . '/.phpoole')) {
-        RecursiveRmdir($path . '/.phpoole');
+function getOptionWebsitePath($opts, $option) {
+    if (isset($opts->$option)) {
+        if (!is_dir($opts->$option)) {
+            echo 'Invalid directory provided' . PHP_EOL
+                . PHP_EOL;
+            echo $opts->getUsageMessage();
+            exit(2);
+        }
+        $websiteDir = $opts->$option;
+        $websitePath = str_replace(DIRECTORY_SEPARATOR, '/', realpath($websiteDir));
+        return $websitePath;
     }
-    mkdir($path . '/.phpoole');
+}
+ 
+function mkInitDir($path, $force = false) {
+    try {
+        if (is_dir($path . '/' . PHPOOLE_DIRNAME)) {
+            if ($force) {
+                RecursiveRmdir($path . '/' . PHPOOLE_DIRNAME);
+            }
+            else {
+                throw new Exception(sprintf('%s directory already exist', PHPOOLE_DIRNAME));
+            }
+        }
+        if (false === mkdir($path . '/' . PHPOOLE_DIRNAME)) {
+            throw new Exception(sprintf('%s not created', $path));
+        }
+        printf('[OK] create %s' . PHP_EOL, PHPOOLE_DIRNAME);
+    }  
+    catch (Exception $e) {
+        printf('[KO] %s' . PHP_EOL, $e->getMessage());
+        exit(2);
+    }
 }
 
-function MakeConfigFile($filePath) {
+function mkConfigFile($filePath) {
     $content = <<<'EOT'
 [site]
-title        = "PHPoole"
-baseline     = "PHPoole is a simple static website/weblog generator written in PHP."
-description  = "PHPoole is (will be!) a simple static website/weblog generator written in PHP. It parses your content written with Markdown, merge it with layouts and generates static HTML files."
-base_url     = "http://localhost/PHPoole"
-language     = "en_US"
+title       = "PHPoole"
+baseline    = "PHPoole is a simple static website generator."
+description = "PHPoole is a simple static website/weblog generator written in PHP. It parses your content written with Markdown, merge it with layouts and generates static HTML files."
+base_url    = "http://localhost:8000"
+language    = "en_US"
 [author]
 name  = "Arnaud Ligny"
 email = "arnaud+phpoole@ligny.org"
 home  = "http://narno.org"
 EOT;
-    if (is_file($filePath)) {
-        unlink($filePath);
+    try {
+        if (false === file_put_contents($filePath, $content)) {
+            throw new Exception(sprintf('%s not created', basename($filePath)));
+        }
+        printf('[OK] create %s/%s' . PHP_EOL, PHPOOLE_DIRNAME, basename($filePath));
     }
-    file_put_contents($filePath, $content);
+    catch (Exception $e) {
+        printf('[KO] %s' . PHP_EOL, $e->getMessage());
+        exit(2);
+    }
 }
 
-function MakeLayoutBaseFile($filePath) {
+function mkLayoutsDir($path) {
+    try {
+        if (false === mkdir(sprintf('%s/layouts', $path))) {
+            throw new Exception(sprintf('%s/layouts not created', $path));
+        }
+        printf('[OK] create %s/layouts' . PHP_EOL, PHPOOLE_DIRNAME);
+    }
+    catch (Exception $e) {
+        printf('[KO] %s' . PHP_EOL, $e->getMessage());
+        exit(2);
+    }
+}
+
+function mkLayoutBaseFile($path, $filename) {
+    $subdir = 'layouts';
     $content = <<<'EOT'
 <!DOCTYPE html>
 <!--[if IE 8]><html class="no-js lt-ie9" lang="en"><![endif]-->
@@ -277,26 +281,61 @@ function MakeLayoutBaseFile($filePath) {
 </body>
 </html>
 EOT;
-    if (is_file($filePath)) {
-        unlink($filePath);
+    try {
+        if (false === file_put_contents("$path/$subdir/$filename", $content)) {
+            throw new Exception(sprintf('%s/%s not created', $subdir, basename($filename)));
+        }
+        printf('[OK] create %s/%s/%s' . PHP_EOL, PHPOOLE_DIRNAME, $subdir, $filename);
     }
-    file_put_contents($filePath, $content);
+    catch (Exception $e) {
+        printf('[KO] %s' . PHP_EOL, $e->getMessage());
+        exit(2);
+    }
 }
 
-function MakeAssetsDir($path) {
-    mkdir($path . '/assets');
-    mkdir($path . '/assets/css');
-    mkdir($path . '/assets/img');
-    mkdir($path . '/assets/js');
+function mkAssetsDir($path) {
+    $subDirList = array(
+        'assets',
+        'assets/css',
+        'assets/img',
+        'assets/js',
+    );
+    try {
+        foreach ($subDirList as $subDir) {
+            if (false === mkdir(sprintf('%s/%s', $path, $subDir))) {
+                throw new Exception(sprintf('%s/%s not created', $path, $subDir));
+            }
+            printf('[OK] create %s/%s' . PHP_EOL, PHPOOLE_DIRNAME, $subDir);
+        }
+    }
+    catch (Exception $e) {
+        printf('[KO] %s' . PHP_EOL, $e->getMessage());
+        exit(2);
+    }
 }
 
-function MakeContentDir($path) {
-    mkdir($path . '/content');
-    mkdir($path . '/content/posts');
-    mkdir($path . '/content/pages');
+function mkContentDir($path) {
+    $subDirList = array(
+        'content',
+        'content/posts',
+        'content/pages',
+    );
+    try {
+        foreach ($subDirList as $subDir) {
+            if (false === mkdir(sprintf('%s/%s', $path, $subDir))) {
+                throw new Exception(sprintf('%s/%s not created', $path, $subDir));
+            }
+            printf('[OK] create %s/%s' . PHP_EOL, PHPOOLE_DIRNAME, $subDir);
+        }
+    }
+    catch (Exception $e) {
+        printf('[KO] %s' . PHP_EOL, $e->getMessage());
+        exit(2);
+    }
 }
 
-function MakeIndexFile($filePath) {
+function mkContentIndexFile($path, $filename) {
+    $subdir = 'content/pages';
     $content = <<<'EOT'
 <!--
 title = PHPoole static website
@@ -310,10 +349,47 @@ It parses your content written with Markdown, merge it with layouts and generate
 
 PHPoole = [PHP](http://www.php.net) + [Poole](http://en.wikipedia.org/wiki/Strange_Case_of_Dr_Jekyll_and_Mr_Hyde#Mr._Poole)
 EOT;
-    if (is_file($filePath)) {
-        unlink($filePath);
+    try {
+        if (false === file_put_contents("$path/$subdir/$filename", $content)) {
+            throw new Exception(sprintf('%s/%s not created', $subdir, basename($filename)));
+        }
+        printf('[OK] create %s/%s/%s' . PHP_EOL, PHPOOLE_DIRNAME, $subdir, $filename);
     }
-    file_put_contents($filePath, $content);
+    catch (Exception $e) {
+        printf('[KO] %s' . PHP_EOL, $e->getMessage());
+        exit(2);
+    }
+}
+
+function mkRouterFile($filePath) {
+    $content = <<<'EOT'
+<?php
+date_default_timezone_set("UTC");
+define("DIRECTORY_INDEX", "index.html");
+$path = parse_url($_SERVER["REQUEST_URI"], PHP_URL_PATH);
+$ext = pathinfo($path, PATHINFO_EXTENSION);
+if (empty($ext)) {
+    $path = rtrim($path, "/") . "/" . DIRECTORY_INDEX;
+}
+if (file_exists($_SERVER["DOCUMENT_ROOT"] . $path)) {
+    return false;
+}
+http_response_code(404);
+echo "404, page not found";
+EOT;
+    try {
+        if (is_file($filePath)) {
+            unlink($filePath);
+        }
+        if (false === file_put_contents($filePath, $content)) {
+            throw new Exception(sprintf('%s not created', basename($filename)));
+        }
+        printf('[OK] create %s/%s' . PHP_EOL, PHPOOLE_DIRNAME, basename($filePath));
+    }
+    catch (Exception $e) {
+        printf('[KO] %s' . PHP_EOL, $e->getMessage());
+        exit(2);
+    }
 }
 
 function parseContent($content, $filename) {
@@ -330,28 +406,6 @@ function parseContent($content, $filename) {
         $info,
         array('body' => $body)
     );
-}
-
-function MakeRouterFile($filePath) {
-    $content = <<<'EOT'
-<?php
-date_default_timezone_set("UTC");
-define("DIRECTORY_INDEX", "index.html");
-$path = parse_url($_SERVER["REQUEST_URI"], PHP_URL_PATH);
-$ext = pathinfo($path, PATHINFO_EXTENSION);
-if (empty($ext)) {
-    $path = rtrim($path, "/") . "/" . DIRECTORY_INDEX;
-}
-if (file_exists($_SERVER["DOCUMENT_ROOT"] . $path)) {
-    return false;
-}
-http_response_code(404);
-echo "404, page not found";
-EOT;
-    if (is_file($filePath)) {
-        unlink($filePath);
-    }
-    file_put_contents($filePath, $content);
 }
